Separate file for focus helpers
